+++ lib/pbio/platform/ev3/pbsysconfig.h
@@ -1,6 +1,9 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2020-2024 The Pybricks Authors
 
+#define PBSYS_CONFIG_APP_BUILTIN_PROGRAMS                        (0)
+#define PBSYS_CONFIG_APP_USER_PROG_FORMAT_MULTI_MPY_V6           (1)
+#define PBSYS_CONFIG_APP_USER_PROG_FORMAT_MULTI_MPY_V6_1_NATIVE  (0)
 #define PBSYS_CONFIG_MAIN                           (1)
 #define PBSYS_CONFIG_STORAGE                        (1)
 #define PBSYS_CONFIG_STORAGE_RAM_SIZE               (10 * 1024)