+++ lib/pbio/sys/hmi.c
@@ -20,12 +20,12 @@
 #include <pbio/event.h>
 #include <pbio/light.h>
 #include <pbsys/config.h>
+#include <pbsys/main.h>
 #include <pbsys/status.h>
 #include <pbsys/storage_settings.h>
 
 #include "light_matrix.h"
 #include "light.h"
-#include "user_program.h"
 
 static struct pt update_program_run_button_wait_state_pt;
 
@@ -52,7 +52,7 @@ static PT_THREAD(update_program_run_button_wait_state(bool button_pressed)) {
 
         // if we made it through a full press and release, without the user
         // program running, then start the user program
-        pbsys_user_program_start_program();
+        pbsys_main_program_request_start(PBSYS_MAIN_PROGRAM_TYPE_USER, 0);
     }
 
     PT_END(pt);